requirements.txt
# Core API framework
fastapi>=0.104.0

# ASGI server (needed to run FastAPI)
uvicorn[standard]>=0.24.0

# Testing framework
pytest>=7.4.0

# HTTP client for testing (used by FastAPI TestClient)
httpx>=0.25.0

# Data validation (already installed, but listing for completeness)
pydantic>=2.0.0

# Environment variable management
python-dotenv>=1.0.0

# Database (optional, planned)
# Not required for the in-memory MVP, but included to support the Postgres migration path.
SQLAlchemy>=2.0.0
alembic>=1.13.0
psycopg[binary]>=3.1.0

# Excel ingestion (RAVE/QC export parsing)
openpyxl>=3.1.0

# Multipart form (file upload for ingest)
python-multipart>=0.0.6

# LLM support (optional, enable with LLM_ENABLED=1)
openai>=1.0.0

# RAG support (embeddings for semantic search)
sentence-transformers>=2.2.0
numpy>=1.24.0
